schema

+++ packages/plugin-processes-api/src/models/FlowCategories.ts
@@ -1,12 +1,20 @@
 import { Model, model } from 'mongoose';
 import * as _ from 'underscore';
 import { IModels } from '../connectionResolver';
-import { IFlowCategory, IFlowCategoryDocument, flowCategorySchema } from './definitions/flowCategories';
+import { FLOW_STATUSES } from './definitions/constants';
+import {
+  IFlowCategory,
+  IFlowCategoryDocument,
+  flowCategorySchema
+} from './definitions/flowCategories';
 
 export interface IFlowCategoryModel extends Model<IFlowCategoryDocument> {
   getFlowCategory(_id: string): Promise<IFlowCategoryDocument>;
   createFlowCategory(doc: IFlowCategory): Promise<IFlowCategoryDocument>;
-  updateFlowCategory(_id: string, doc: IFlowCategory): Promise<IFlowCategoryDocument>;
+  updateFlowCategory(
+    _id: string,
+    doc: IFlowCategory
+  ): Promise<IFlowCategoryDocument>;
   removeFlowCategory(_id: string): void;
 }
 
@@ -31,7 +39,7 @@ export const loadFlowCategoryClass = (models: IModels) => {
     public static async createFlowCategory(doc: IFlowCategory) {
       const flowCategory = await models.FlowCategories.create({
         ...doc,
-        createdAt: new Date(),
+        createdAt: new Date()
       });
 
       return flowCategory;
@@ -41,11 +49,11 @@ export const loadFlowCategoryClass = (models: IModels) => {
      * Update FlowCategory
      */
     public static async updateFlowCategory(_id: string, doc: IFlowCategory) {
-      const flowCategory = await models.FlowCategories.getFlowCategory(_id,);
+      const flowCategory = await models.FlowCategories.getFlowCategory(_id);
 
       await models.FlowCategories.updateOne({ _id }, { $set: { ...doc } });
 
-      const updated = await models.FlowCategories.getFlowCategory( _id );
+      const updated = await models.FlowCategories.getFlowCategory(_id);
 
       return updated;
     }
@@ -55,6 +63,17 @@ export const loadFlowCategoryClass = (models: IModels) => {
      */
     public static async removeFlowCategory(_id: string) {
       await models.FlowCategories.getFlowCategory(_id);
+
+      let count = await models.JobRefers.countDocuments({
+        categoryId: _id,
+        status: { $ne: FLOW_STATUSES.ARCHIVED }
+      });
+      count += await models.JobCategories.countDocuments({ parentId: _id });
+
+      if (count > 0) {
+        throw new Error("Can't remove a flow category");
+      }
+
       return models.FlowCategories.deleteOne({ _id });
     }
   }
